列車種別の選択時ラベルの生成処理のhook化
